add new file for historyfactory
